// Frontend 1.0/src/app/models/pendingRequestModel.ts
export class pendinRequest{
    id: string;
    requestType: string;
    requestMadeDate: string;
    beginDate: string;
    dueDate: string;
    dateDHConfirmed: string;
    dateAMConfirmed: string;
    assetId: string;
    requestedNic: string;
    description: string;
    isAssigned: string;
    isApprovedByAssetManager: string;
    is_approved_by_department_head: string;





}